Determine which M/E bus to use on the ATEM via preference.
There's deliberately no UI to change this preference - we don't really
want people changing it unless they know what they're doing, and being
able to edit a JSON file in `~/.config/av-control` is a nice barrier for
well-meaning interlopers. In any case, we don't support changing the
value at runtime - it's passed to the constructor of VisualsSystem.

This affects:
- Cut/Fade buttons
- Preview buttons
- Send mix to aux / all buttons